refactor(calendar): calendar 専用 store/constants を features/calendar 配下へ移動
PR #1180 で user preference を分離し、PR #1182 で dead state を削除した結果、
useCalendarSettingsStore は純粋に Calendar UI state のみになった。
本来の所在地である features/calendar/ 配下へ物理移動し、当初の PR2 ゴール
（lib/ を feature 非依存の汎用道具に揃える）を完了させる。

仕様変更・UI 変更なし。devtools store name も不変。

移動:
- lib/stores/useCalendarSettingsStore.ts → features/calendar/stores/
- lib/stores/useCalendarNavigationStore.ts → features/calendar/stores/
- lib/stores/userSettings.ts → features/calendar/stores/
- lib/calendar-constants.ts → features/calendar/lib/constants.ts

DAG 違反解消:
- chronotype-settings (L0) → calendar (L2) を解消するため、
  ChronotypeSettings に onChronotypeChange callback prop を追加し、
  profile-settings.tsx (features/settings) から store 更新を注入

barrel 追加:
- features/calendar/index.ts に useCalendarSettingsStore /
  useCalendarNavigationStore / UserSettings / CalendarSettings を公開
  （app/ composition 層が barrel 経由で参照するため）

検証:
- pnpm --filter @dayopt/product typecheck / lint / test:run (1684 pass) / build
- pnpm lint:boundaries (0 violations)
- pnpm --filter @dayopt/storybook build
