+++ apps/product/src/features/calendar/lib/constants.ts
