+++ apps/product/src/features/calendar/types/calendar.types.ts
@@ -2,10 +2,9 @@
 // feature 間で共有するため features/ 内には定義しない
 export type { CalendarEvent } from '@/lib/types/calendar-event';
 
-// CalendarViewType 関連は共有層（@/lib/calendar-constants）が canonical source
-// store も参照するため feature 内には置けない
-export { getMultiDayCount, isMultiDayView } from '@/lib/calendar-constants';
-export type { CalendarViewType, MultiDayCount, MultiDayViewType } from '@/lib/calendar-constants';
+// CalendarViewType 関連は feature 内の lib/constants が canonical source
+export { getMultiDayCount, isMultiDayView } from '../lib/constants';
+export type { CalendarViewType, MultiDayCount, MultiDayViewType } from '../lib/constants';
 
 /** ビューの日付範囲（開始・終了・各日の配列） */
 export interface ViewDateRange {
