+++ apps/product/src/features/settings/components/profile-settings.tsx
@@ -1,15 +1,17 @@
 'use client';
 
-import { useState } from 'react';
+import { useCallback, useState } from 'react';
 
 import { Camera } from 'lucide-react';
 import { useTranslations } from 'next-intl';
 
 import { useAuthStore } from '@/features/auth';
+import { useCalendarSettingsStore } from '@/features/calendar/stores/useCalendarSettingsStore';
 import { ChronotypeSettingsPanel as ChronotypeSettings } from '@/features/chronotype';
 import { LabeledRow } from '@/lib/components/common/LabeledRow';
 import { SectionCard } from '@/lib/components/common/SectionCard';
 import { Avatar, AvatarFallback, AvatarImage } from '@/lib/components/ui/avatar';
+import type { ChronotypeSettings as ChronotypeSettingsState } from '@/lib/types/chronotype';
 import { getAvatarUrl, getDisplayName, getInitials } from '@/lib/user';
 
 import { AvatarChangeDialog } from './avatar-change-dialog';
@@ -23,12 +25,20 @@ import { DisplayNameDialog } from './display-name-dialog';
 export function ProfileSettings() {
   const t = useTranslations();
   const user = useAuthStore((state) => state.user);
+  const updateCalendarSettings = useCalendarSettingsStore((state) => state.updateSettings);
 
   const [showAvatarDialog, setShowAvatarDialog] = useState(false);
   const [showDisplayNameDialog, setShowDisplayNameDialog] = useState(false);
 
   const avatarUrl = getAvatarUrl(user);
   const displayName = getDisplayName(user);
+
+  const handleChronotypeChange = useCallback(
+    (chronotype: ChronotypeSettingsState | null) => {
+      updateCalendarSettings({ chronotype });
+    },
+    [updateCalendarSettings],
+  );
 
   return (
     <div className="space-y-6 sm:space-y-8">
@@ -61,7 +71,7 @@ export function ProfileSettings() {
       </SectionCard>
 
       {/* クロノタイプ */}
-      <ChronotypeSettings />
+      <ChronotypeSettings onChronotypeChange={handleChronotypeChange} />
 
       {/* Dialogs */}
       <AvatarChangeDialog open={showAvatarDialog} onOpenChange={setShowAvatarDialog} />
