+++ apps/product/src/features/settings/stores/useCalendarSettingsStore.test.ts
@@ -1,6 +1,6 @@
 import { beforeEach, describe, expect, it } from 'vitest';
 
-import { useCalendarSettingsStore } from '@/lib/stores/useCalendarSettingsStore';
+import { useCalendarSettingsStore } from '@/features/calendar/stores/useCalendarSettingsStore';
 
 describe('useCalendarSettingsStore', () => {
   beforeEach(() => {
